// src/models/IParkWorkVehicleModel.ts
export interface IParkWorkVehicleModel {
  vehicleNumber?: number;
  numberOfMotorVehiclesInTransportMvit?: number;
  numberOfOccupants?: number;
  dayOfCrash?: number;
  monthOfCrash?: number;
  hourOfCrash?: number;
  minuteOfCrash?: number;
  firstHarmfulEvent?: number;
  firstHarmfulEventName?: string;
  mannerOfCollision?: number;
  mannerOfCollisionName?: string;
  unitType?: string;
  hitAndRun?: string;
  registrationState?: number;
  registrationStateName?: string;
  registeredVehicleOwner?: number;
  registeredVehicleOwnerName?: string;
  vehicleMake?: number;
  vehicleMakeName?: string;
  vehicleModel?: number;
  makeModelCombined?: number;
  bodyType?: number;
  bodyTypeName?: string;
  vehicleModelYear?: number;
  vehicleIdentificationNumberVin?: string;
  vehicleTrailing?: string;
  mcidIssuingAuthority?: number;
  mcidIssuingAuthorityName?: string;
  mcidIdentificationNumber?: string;
  motorCarrierIdentificationNumber?: string;
  grossVehicleWeightRating?: number;
  vehicleConfiguration?: number;
  cargoBodyType?: number;
  hazardousMaterialInvolvement?: string;
  hazardousMaterialPlacard?: string;
  hazardousMaterialIdentificationNumber?: number;
  hazardousMaterialClassNumber?: number;
  releaseOfHazardousMaterialFromTheCargoCompartment?: string;
  busUse?: string;
  specialUse?: number;
  specialUseName?: string;
  emergencyMotorVehicleUse?: string;
  underrideOverride?: number;
  underrideOverrideName?: string;
  initialContactPoint?: number;
  extentOfDamage?: string;
  vehicleRemoval?: string;
  mostHarmfulEvent?: number;
  relatedFactorsVehicleLevel1?: number;
  relatedFactorsVehicleLevel2?: number;
  fireOccurrence?: string;
  fatalitiesInVehicle?: number;
}